WIP extended selector parsing
