+++ extended-css/src/parse.ts
@@ -2,22 +2,10 @@ import * as CSSTree from 'css-tree';
 
 import { Child } from './combinators/child';
 import { Descendant } from './combinators/descendant';
-import { MatchesPath } from './extendedPseudoClasses';
+import { Contains, MatchesPath } from './extendedPseudoClasses';
 import { Mixed } from './mixed';
 import { RawQuery } from './raw';
 import { Selector } from './types';
-
-export function select(selectors: Selector[]) {
-  let nodes: Element[] = [document.body];
-  for (const selector of selectors) {
-    nodes = selector.select(nodes);
-    console.log(selector, nodes);
-    if (nodes.length === 0) {
-      return nodes;
-    }
-  }
-  return nodes;
-}
 
 export function parse(rule: string): Selector[] {
   const ast = CSSTree.parse(rule, { context: 'selector', positions: true });
@@ -25,8 +13,6 @@ export function parse(rule: string): Selector[] {
   const res: Selector[] = [];
 
   CSSTree.walk(ast, (node) => {
-    const literal = rule.slice(node.loc!.start.offset, node.loc!.end.offset);
-
     let selector: Selector;
     let action;
     switch (node.type) {
@@ -34,14 +20,24 @@ export function parse(rule: string): Selector[] {
         return;
       case 'IdSelector':
       case 'ClassSelector':
-        if (res.length > 0) {
-          selector = new Mixed(literal);
+      case 'TypeSelector':
+      case 'AttributeSelector': {
+        const literal = rule.slice(node.loc!.start.offset, node.loc!.end.offset);
+        if (res.length === 0) {
+          res.push(new RawQuery(literal));
+        } else if (res[res.length - 1] instanceof Descendant) {
+          res[res.length - 1] = new RawQuery(literal);
         } else {
-          selector = new RawQuery(literal);
+          res.push(new Mixed(literal));
+        }
+
+        if (node.type === 'AttributeSelector') {
+          action = CSSTree.walk.skip;
         }
         break;
+      }
       case 'Combinator':
-        selector = parseCombinator(node.name);
+        res.push(parseCombinator(node.name));
         break;
       case 'PseudoClassSelector': {
         let skip;
@@ -49,12 +45,12 @@ export function parse(rule: string): Selector[] {
         if (skip) {
           action = CSSTree.walk.skip;
         }
+        res.push(selector);
         break;
       }
       default:
         throw new Error(`Unsupported node type ${node.type}`);
     }
-    res.push(selector);
     if (action) {
       return action;
     }
@@ -82,6 +78,13 @@ function parsePseudoClass(node: CSSTree.PseudoClassSelector): { selector: Select
         throw new Error('Bad matches-path');
       }
       return { selector: new MatchesPath(child.value), skip: true };
+    }
+    case 'contains': {
+      const child = node.children?.first;
+      if (child?.type !== 'Raw') {
+        throw new Error('Bad contains');
+      }
+      return { selector: new Contains(child.value), skip: true };
     }
     default:
       throw new Error('Unsupported pseudoclass');
